+++ guicast/bcpan.C
@@ -174,7 +174,7 @@ int BC_Pan::button_release_event()
 	return 0;
 }
 
-void BC_Pan::repeat_event(int64_t duration)
+void BC_Pan::repeat_event(int duration)
 {
 	if(duration == top_level->get_resources()->tooltip_delay &&
 		tooltip_text[0] != 0 &&
