+++ guicast/bcslider.C
@@ -172,7 +172,7 @@ void BC_Slider::show_value_tooltip()
 }
 
 
-void BC_Slider::repeat_event(int64_t duration)
+void BC_Slider::repeat_event(int duration)
 {
 	if(duration == top_level->get_resources()->tooltip_delay)
 	{
